User routes moved into users routing file
